% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.mix.MCMC.R
\name{get.nrmlzd.sample.covariance}
\alias{get.nrmlzd.sample.covariance}
\title{Converts allele count and sample size data to a normalized 
allele frequency covariance matrix}
\usage{
get.nrmlzd.sample.covariance(counts, sample.sizes)
}
\arguments{
\item{counts}{A matrix of allele counts for which each row corresponds to a 
sample and each column corresponds to a locus. The \eqn{ij}th entry gives the number 
 of times the counted allele at the \eqn{j}th locus is observed in the \eqn{i}th sample.
 Please see the vignette for a discussion of what the allele count data matrix 
should look like.}

\item{sample.sizes}{A matrix of sample sizes for which each row corresponds to a 
 sample and each column corresponds to a locus. The \eqn{ij}th entry gives the number 
of chromosomes genotyped in the \eqn{i}th sample at the \eqn{j}th locus.}
}
\value{
This function returns a list with three components:
\itemize{
\item norm.sample.covariance - normalized sample allele frequency covariance matrix.
\item mean.sample.sizes - The vector of length \eqn{K} giving the mean sample size  
across loci in each of the \eqn{K} samples.
\item loci - The number of loci in the dataset.
}
}
\description{
This function generates a normalized covariance matrix from 
allele count and sample size data.  The output can be used 
to run a SpaceMix analysis using the \code{run.spacemix.analysis}
function. Running a SpaceMix analysis on this normalized covariance 
matrix (rather than on raw allele count and sample size data, or 
on sample allele frequency data) is recommended when there are many 
loci for which there are missing data, as its use results in pairwise, 
rather than listwise, deletion of missing data.
}
\details{
The normalization is used to standardize the variance in 
allele frequencies across loci, and is performed by dividing 
the allele frequencies at a locus by 
\eqn{\sqrt{\bar{f}(1-\bar{f})}} where \eqn{\bar{f}} 
is the average of the \eqn{K} sample allele frequencies,
weighted by mean population size.  That is, 
\deqn{\bar{f}_{\ell} = \frac{1}{\sum_K S_{k,\ell}} 
\sum_K \hat{f}_{k,\ell} S_{k,\ell}}.
}
\examples{
# load example dataset
data(spacemix.example.dataset)

# generate normalized sample allele frequency covariance
norm.samp.cov <- get.nrmlzd.sample.covariance(counts = spacemix.example.dataset$allele.counts,
				sample.sizes = spacemix.example.dataset$sample.sizes)

# visualize results
image(norm.samp.cov$norm.sample.covariance,
		  main="Normalized sample allele frequency covariance",
		  xaxt='n',yaxt='n',xlab="Population",ylab="Population")
axis(side=1,c(0:29)/29,labels=c(paste(1:30)))
axis(side=2,c(0:29)/29,labels=c(paste(1:30)),las=2)
}
