% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/space.mix.MCMC.R
\name{make.spacemix.map}
\alias{make.spacemix.map}
\title{Plots the output of a SpaceMix analysis}
\usage{
make.spacemix.map(spacemix.map.list, text = FALSE, ellipses = TRUE,
  source.option = TRUE, xlim = NULL, ylim = NULL)
}
\arguments{
\item{spacemix.map.list}{This is the name of the list produced using the function
\code{\link{make.spacemix.map.list}}.  This list should already be loaded into
R's working memory.}

\item{text}{This option (\code{TRUE} or \code{FALSE}) specifies whether sample
names should be printed on the plotted figure at the locations of the MAP
parameter estimates for each sample's geogenetic location and that of its
admixture source location.}

\item{ellipses}{This option (\code{TRUE} or \code{FALSE}) specifies whether
bivariate credible interval ellipses should be plotted for sample geogenetic
locations and admixture source locations.}

\item{source.option}{This option (\code{TRUE} or \code{FALSE}) specifies whether
the sources of admixture for each sample will be plotted, with accompanying
admixture arrows.}

\item{xlim}{This vector of length 2 gives the minimum and maximum x-value of the
plotting window.  Default is NULL, in which case the \code{xlim} used will
be the minimum and maximum credible ellipse x-coordinates.}

\item{ylim}{This vector of length 2 gives the minimum and maximum y-value of the
        plotting window. Default is NULL, in which case the \code{xlim} used will
        be the minimum and maximum credible ellipse y-coordinates.}
}
\value{
This function plots the output of a SpaceMix analysis.  If specified,
		sample names are plotted at the MAP estimates of their geogenetic and admixture
		source locations.  Ellipses with solid margins denote credible intervals on
		the posterior distributions of sample geogenetic locations, while those with dashed
		margins denote credible intervals on the posterior distributions of sample admixture
		source locations. The return value of this function is invisible.
}
\description{
This function plots the output of a SpaceMix analysis. Users can specify
whether they wish to also visualize the names of the samples, the
credible ellipses of estimated sample geogenetic and admixture source
location coordinates, as well as whether they wish to visualize the
sources of admixture at all.
}
\examples{
#load example spacemix.map.list object
data(example.spacemix.map.list)

#make SpaceMix output map
#	without plotting admixture sources, and with sample names plotted
make.spacemix.map(example.spacemix.map.list,source=FALSE,text=TRUE)
}

