% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MCN.frequencies.list}
\alias{MCN.frequencies.list}
\title{Example model adequacy dataset}
\format{A list with 5 elements
\describe{
  \item{mean.sample.sizes}{
		vector of mean sample sizes for all samples
	with \code{length} = number of samples}
  \item{sample.frequencies}{
	matrix of sample frequencies 
	(counts/sample.sizes) with 
	\code{nrow} = number of samples and 
	\code{ncol} = number of loci.}
  \item{normalized.sample.frequencies}{
	matrix of sample frequencies 
	normalized by mean.f * (1-mean.f),
	where mean.f at a locus is the 
	mean alelle frequency at that locus}
  \item{mean.centered.sample.frequencies}{
	matrix of sample frequencies with the 
	mean allele frequency at each locus 
	subtracted from all entries at that locus}
  \item{mean.centered.normalized.sample.frequencies}{
	matrix of sample frequencies that is 
	both normalized and mean-centered as described above.}
}}
\usage{
MCN.frequencies.list
}
\description{
Example output of processed (mean-centered 
and normalized) allele frequency data produced 
by running a SpaceMix analysis
}
\keyword{datasets}
