% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/space.mix.MCMC.R
\name{plot.admix.arrows}
\alias{plot.admix.arrows}
\title{Plots admixture arrows}
\usage{
\method{plot}{admix.arrows}(admix.source.coords, geogen.coords,
  admix.proportions, colors = NULL, length = 0.1)
}
\arguments{
\item{admix.source.coords}{This is a k x 2 matrix of coordinates in which
the two elements of the ith row give the x-coordinate and y-coordinate
of the source admixture for sample i.}

\item{geogen.coords}{This is a k x 2 matrix of coordinates in which
the two elements of the ith row give the x-coordinate and y-coordinate
of the geogenetic location of the ith sample.}

\item{admix.proportions}{This is a vector of length k in which each element
is the estimated admixture proportion in the corresponding sample.}

\item{colors}{This is a vector of colors of length k in which each element
gives the color in which the corresponding sample should be plotted.}

\item{length}{This value determines the length of the edges of the arrow head
        (in inches).  The default value is 0.1.}
}
\value{
This function plots the admixture arrows on an existing plotting window.
		Its return value is invisible.
}
\description{
This function plots the admixture arrows, from the sources of admixture to their
targets, on the geogenetic map estimated as part of a SpaceMix analysis.
The width and opacity of these arrows can be proportional to their admixture
proportions, which provides an intuitive visualization of the amount of admixture
each sample is drawing.
}

