% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spacemix.example.dataset}
\alias{spacemix.example.dataset}
\title{Example SpaceMix dataset}
\format{A list with 3 elements
\describe{
  \item{allele.counts}{
		matrix of allele counts with
		\code{nrow} = number of samples and
		\code{ncol} = number of loci}
  \item{sample.sizes}{
		matrix of sample sizes with
		\code{nrow} = number of samples and
		\code{ncol} = number of loci}
  \item{population.coordinates}{
		matrix of longitude and latitude with
		\code{nrow} = number of samples and
		\code{ncol} = 2}
}}
\usage{
spacemix.example.dataset
}
\description{
Example genetic data and geographic metadata
for running a SpaceMix analysis
}
\keyword{datasets}

