% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spacemix.location.data}
\alias{spacemix.location.data}
\title{Example location data}
\format{A list with 5 elements, 
	where \code{K} = the number of samples
\describe{
  \item{geogen.coords}{
		a matrix (2 x \code{K}) of geogenetic location coordinates 
	generated for each sample in a SpaceMix run}
  \item{admix.coords}{
		a matrix (2 x \code{K}) of admixture source location 
	coordinates generated for each sample in a SpaceMix run}
  \item{sample.coords}{
		a matrix (\code{K} x 2) of geographic sampling  
	coordinates for the samples in the dataset}
  \item{pop.colors}{
	a vector (\code{length = K}) of colors for 
	pretty plotting of sample coordinates}
  \item{admix.colors}{
	a vector (\code{length = K}) of colors for 
	admixture sources, generated using 
	\code{fade.admixture.source.points},
	which makes the opacity of the \code{pop.colors} 
	proportional to the admixture proportion for each sample}
}}
\usage{
spacemix.location.data
}
\description{
Example location data for visualizing
how a Procrustes superimposition works
}
\keyword{datasets}
